requirements.txt
fastapi
sqlalchemy
uvicorn
orjson
black
isort
pydantic-settings
asyncpg
psycopg
psycopg[binary]
passlib[bcrypt]
python-jose
openai
APScheduler